requirements.txt
Flask>=2.0.1
requests>=2.26.0
python-magic>=0.4.24
Werkzeug>=2.0.1
Flask-CORS>=3.0.10
python-dotenv>=0.19.0
gunicorn>=20.1.0 